+++ src/main/java/gtb/api/unification/materials/GTBMaterialFlagAddition.java
@@ -1,4 +1,4 @@
-package gtb.api.unification;
+package gtb.api.unification.materials;
 
 import org.jetbrains.annotations.ApiStatus;
